requirements.txt
influxdb-client
pyyaml
schema
flask
waitress
jq
# influxdb